// src/app/models/TableConfigure.ts
export class TableConfigure{
    name: string | undefined;
    explicitName: string | undefined;
    isHidden: boolean | undefined;
    actionGroup: string | undefined;
    columns: null | undefined;
    id: number | undefined;
    createdDate: Date | undefined ;
    modifiedDate: Date | undefined ;
    isDeleted:boolean | undefined;
}